Move RoomManager classes to room package
