+++ app/src/internal/java/com/twilio/video/app/VideoApplicationComponent.java
@@ -22,10 +22,10 @@ import com.twilio.video.app.data.api.VideoAppServiceModule;
 import com.twilio.video.app.ui.ScreenSelectorModule;
 import com.twilio.video.app.ui.login.LoginActivityModule;
 import com.twilio.video.app.ui.room.RoomActivityModule;
+import com.twilio.video.app.ui.room.RoomManagerModule;
 import com.twilio.video.app.ui.settings.SettingsActivityModule;
 import com.twilio.video.app.ui.settings.SettingsFragmentModule;
 import com.twilio.video.app.ui.splash.SplashActivityModule;
-import com.twilio.video.app.videosdk.RoomManagerModule;
 import dagger.Component;
 import dagger.android.AndroidInjectionModule;
