+++ app/src/main/java/com/twilio/video/app/ui/room/RoomActivity.java
@@ -66,24 +66,19 @@ import com.twilio.video.app.R;
 import com.twilio.video.app.adapter.StatsListAdapter;
 import com.twilio.video.app.base.BaseActivity;
 import com.twilio.video.app.data.Preferences;
+import com.twilio.video.app.ui.room.RoomViewEffect.RequestScreenSharePermission;
+import com.twilio.video.app.ui.room.RoomViewEffect.ScreenShareError;
+import com.twilio.video.app.ui.room.RoomViewEvent.ConnectToRoom;
+import com.twilio.video.app.ui.room.RoomViewEvent.DisconnectFromRoom;
+import com.twilio.video.app.ui.room.RoomViewEvent.SetupLocalMedia;
+import com.twilio.video.app.ui.room.RoomViewEvent.StartScreenCapture;
+import com.twilio.video.app.ui.room.RoomViewEvent.StopScreenCapture;
+import com.twilio.video.app.ui.room.RoomViewEvent.TearDownLocalMedia;
+import com.twilio.video.app.ui.room.RoomViewEvent.ToggleLocalAudio;
+import com.twilio.video.app.ui.room.RoomViewEvent.ToggleSpeakerPhone;
 import com.twilio.video.app.ui.settings.SettingsActivity;
 import com.twilio.video.app.util.InputUtils;
 import com.twilio.video.app.util.StatsScheduler;
-import com.twilio.video.app.videosdk.ParticipantViewState;
-import com.twilio.video.app.videosdk.RoomManager;
-import com.twilio.video.app.videosdk.RoomViewEffect;
-import com.twilio.video.app.videosdk.RoomViewEffect.RequestScreenSharePermission;
-import com.twilio.video.app.videosdk.RoomViewEffect.ScreenShareError;
-import com.twilio.video.app.videosdk.RoomViewEvent;
-import com.twilio.video.app.videosdk.RoomViewEvent.ConnectToRoom;
-import com.twilio.video.app.videosdk.RoomViewEvent.DisconnectFromRoom;
-import com.twilio.video.app.videosdk.RoomViewEvent.SetupLocalMedia;
-import com.twilio.video.app.videosdk.RoomViewEvent.StartScreenCapture;
-import com.twilio.video.app.videosdk.RoomViewEvent.StopScreenCapture;
-import com.twilio.video.app.videosdk.RoomViewEvent.TearDownLocalMedia;
-import com.twilio.video.app.videosdk.RoomViewEvent.ToggleLocalAudio;
-import com.twilio.video.app.videosdk.RoomViewEvent.ToggleSpeakerPhone;
-import com.twilio.video.app.videosdk.RoomViewState;
 import java.util.List;
 import javax.inject.Inject;
 import org.jetbrains.annotations.NotNull;
