+++ app/src/main/java/com/twilio/video/app/ui/room/RoomViewEffect.kt
@@ -1,4 +1,4 @@
-package com.twilio.video.app.videosdk
+package com.twilio.video.app.ui.room
 
 sealed class RoomViewEffect {
     object ScreenShareError : RoomViewEffect()
